+++ main.go
@@ -1,5 +1,6 @@
 package main
 
+import "github.com/cliffano/packer-post-processor-json-updater/json-updater"
 import "github.com/hashicorp/packer/packer/plugin"
 
 func main() {
@@ -7,6 +8,6 @@ func main() {
 	if err != nil {
 		panic(err)
 	}
-	server.RegisterPostProcessor(&PostProcessor{})
+	server.RegisterPostProcessor(new(json-updater.PostProcessor))
 	server.Serve()
 }